feature: rename pyasic_dashboard to pyasic_tools
