+++ pyproject.toml
@@ -1,9 +1,9 @@
 [tool.poetry]
-name = "pyasic_dashboard"
+name = "pyasic_tools"
 version = "0.1.0"
 description = "Dashboard for visualizing pyasic data"
 authors = ["Wilfred Allyn <wilfredallyn@protonmail.com>"]
-repository = "https://github.com/wilfredallyn/pyasic_dashboard"
+repository = "https://github.com/wilfredallyn/pyasic_tools"
 license = "MIT"
 readme = "README.md"
