+++ tests/__init__.py
@@ -4,5 +4,5 @@ from tests.db_tests import DBTest
 
 if __name__ == "__main__":
     # python -m unittest discover
-    # `coverage run --source pyasic_dashboard -m unittest discover` will give code coverage data
+    # `coverage run --source pyasic_tools -m unittest discover` will give code coverage data
     unittest.main()
